✨ feat: crud-delet

+++ listar-usuario.php
@@ -17,7 +17,7 @@
             print "<hr>";
             print "<a class='btn btn-success' href='?page=editar&id=".$row->id."'>Editar</a>";
             print " | ";
-            print "<a class='btn btn-danger' href='?page=excluir&id=$row->id'>Excluir</a>";
+            print "<a class='btn btn-danger' onclick=\"if(confirm('Tem certeza que deseja excluir?')){location.href='?page=salvar&acao=excluir&id=".$row->id."'}else{false;}\">Excluir</a>";
             print "<hr>";
         }
     }else{
